% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_for_two_levels}
\alias{check_for_two_levels}
\title{Checks whether a vector has more than two unique values}
\usage{
check_for_two_levels(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Checks whether a vector has more than two unique values
}
\keyword{internal}
