% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{compute_smd_parts}
\alias{compute_smd_parts}
\title{Compute components of SMD}
\usage{
compute_smd_parts(.x, .g, .w, applyFUN = n_mean_var)
}
\arguments{
\item{.x}{a vector of values}

\item{.g}{a vector of groupings to compare}

\item{.w}{a vector of weights (optional)}

\item{applyFUN}{the \code{FUN} used to compute the SMD parts. Defaults to
\code{\link{n_mean_var}}}
}
\description{
Computes \code{D} and \code{S} for use within \link{compute_smd}.
}
