% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{compute_smd_var}
\alias{compute_smd_var}
\title{Computes SMD variance}
\usage{
compute_smd_var(d, smd_parts)
}
\arguments{
\item{d}{an SMD value}

\item{smd_parts}{a \code{list} of components for from \code{\link{compute_smd_parts}}
computing standardized mean differences}
}
\description{
Calculates the variance of a standardized mean difference using the method of
Hedges and Olkin (1985):
}
\details{
\deqn{
\sqrt{\frac{n_1 + n_2}{n_1n_2} + \frac{d^2}{2(n_1 + n_2)}}
}
}
\keyword{internal}
