\name{MultDiffModel-class}
\docType{class}
\alias{MultDiffModel-class}

\title{Class "MultDiffModel"}
\description{

  \code{MultDiffModel} is a virtual class which represents a
  multivariate diffusion model.
  
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{data}:}{Object of class \code{"ContinuousProcess"}}
  }
}
\section{Methods}{
No methods defined with class "MultDiffModel" in the signature.
}
\author{
Nina Munkholt, \email{nina.m@math.ku.dk}
}
\examples{
showClass("MultDiffModel")
}
\keyword{classes}
