% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_var.R
\docType{methods}
\name{n_mean_var}
\alias{n_mean_var}
\alias{n_mean_var,numeric-method}
\alias{n_mean_var,logical-method}
\alias{n_mean_var,factor-method}
\alias{n_mean_var,character-method}
\title{Compute n, mean and variance}
\usage{
n_mean_var(x)

\S4method{n_mean_var}{numeric}(x)

\S4method{n_mean_var}{logical}(x)

\S4method{n_mean_var}{factor}(x)

\S4method{n_mean_var}{character}(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a list containing \code{mean} and \code{var}
}
\description{
Available for \code{numeric}, \code{integer}, and \code{factor} objects.
\code{character} objects are handled by first converting to a factor.
}
