% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_var.R
\name{proportions}
\alias{proportions}
\title{Computes proportions in each factor level}
\usage{
proportions(x)
}
\arguments{
\item{x}{a factor variable. It must be \code{sort}ed first!}
}
\description{
Computes proportions in each factor level
}
