% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd}
\alias{smd}
\title{Compute Standardized Mean Difference}
\usage{
smd(x, g, method = NA_character_)
}
\arguments{
\item{x}{a vector of values}

\item{g}{a vector}

\item{method}{not yet implemented}
}
\value{
the standardized mean differences between levels of \code{g}
for values of \code{x}
}
\description{
Computes the standardized mean differnce (SMD) between two groups.
}
\seealso{
\code{\link{compute_smd}} for mathematical details
}
