% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd_tidier}
\alias{smd_tidier}
\alias{tidy_smd_singlevar}
\alias{tidy_smd_multiplevar}
\title{Helper to clean up smd output}
\usage{
tidy_smd_singlevar(smd_res)

tidy_smd_multiplevar(smd_res)
}
\arguments{
\item{smd_res}{a result of \link{smd}}
}
\value{
a \code{data.frame}
}
\description{
Helper to clean up smd output
}
\keyword{internal}
