% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_var.R
\name{variance_computations}
\alias{variance_computations}
\alias{multinom_var}
\title{Variance computations}
\usage{
multinom_var(p)
}
\arguments{
\item{p}{a vector of proportions corresponding to the proportion in each group}
}
\value{
a covariance matrix
}
\description{
Variance computations
}
\keyword{internal}
