% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{compute_smd_parts}
\alias{compute_smd_parts}
\title{Compute components of SMD}
\usage{
compute_smd_parts(x, g, tapplyMethod = tapply, tapplyFUN = mean_var,
  tapplyArgs = list())
}
\arguments{
\item{x}{a vector of values}

\item{g}{a vector of groupings to compare}

\item{tapplyMethod}{either \code{tapply} or \code{\link[fastmatch]{ctapply}}.
Defaults to \code{\link[fastmatch]{ctapply}}.}

\item{tapplyFUN}{the \code{FUN} argument passed to \code{tapplyMethod}}

\item{tapplyArgs}{a \code{list} of arguments passed to \code{tapplyMethod}}
}
\description{
Computes \code{D} and \code{S} for use within \link{compute_smd}.
}
