% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_var.R
\name{n_mean_var}
\alias{n_mean_var}
\alias{n_mean_var,numeric,missing-method}
\alias{n_mean_var,numeric,numeric-method}
\alias{n_mean_var,integer,missing-method}
\alias{n_mean_var,integer,numeric-method}
\alias{n_mean_var,logical,missing-method}
\alias{n_mean_var,logical,numeric-method}
\alias{n_mean_var,factor,missing-method}
\alias{n_mean_var,factor,numeric-method}
\alias{n_mean_var,character,missing-method}
\alias{n_mean_var,character,numeric-method}
\title{Compute n, mean and variance}
\usage{
n_mean_var(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{numeric,missing}(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{numeric,numeric}(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{integer,missing}(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{integer,numeric}(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{logical,missing}(x, na.rm = FALSE)

\S4method{n_mean_var}{logical,numeric}(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{factor,missing}(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{factor,numeric}(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{character,missing}(x, w = NULL, na.rm = FALSE)

\S4method{n_mean_var}{character,numeric}(x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{a vector of values}

\item{w}{an optional vector of \code{numeric} weights}

\item{na.rm}{passed to \code{sum}}
}
\value{
a list containing \code{mean} and \code{var}
}
\description{
Available for \code{numeric}, \code{integer}, and \code{factor} objects.
\code{character} objects are handled by first converting to a factor.
}
\keyword{internal}
