\name{smd-package}
\alias{smd-package}
\alias{smd}
\docType{package}
\title{Sparse Multivariate Diffusions}
\description{
  
  Estimation of high dimensional, discretely observed
  diffusions using a strategy which yields a sparse representation of
  the drift term.
  
}
\details{
\tabular{ll}{
Package: \tab smd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-08-04\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
  Nina Munkholt

Maintainer: Nina Munkholt \email{nina.m@math.ku.dk}
}
\keyword{ package }

