% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\docType{methods}
\name{smd}
\alias{smd}
\alias{smd,logical-method}
\alias{smd,matrix-method}
\alias{smd,list-method}
\alias{smd,data.frame-method}
\title{Compute Standardized Mean Difference}
\usage{
smd(x, g, method = NA_character_)

\S4method{smd}{logical}(x, g, method = NA_character_)

\S4method{smd}{matrix}(x, g, method = NA_character_)

\S4method{smd}{list}(x, g, method = NA_character_)

\S4method{smd}{data.frame}(x, g, method = NA_character_)
}
\arguments{
\item{x}{a vector of values}

\item{g}{a vector of groupings to compare}

\item{method}{not yet implemented}
}
\value{
the standardized mean differences between levels of \code{g}
for values of \code{x}
}
\description{
Computes the standardized mean differnce (SMD) between two groups.
}
\examples{
x <- rnorm(100)
g <- rep(1:2, each = 50)
smd(x, g)
}
\seealso{
\code{\link{compute_smd}} for mathematical details
}
