% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\docType{methods}
\name{smd}
\alias{smd}
\alias{smd,numeric-method}
\alias{smd,integer-method}
\alias{smd,logical-method}
\alias{smd,factor-method}
\title{Compute Standardized Mean Difference}
\usage{
smd(x, g, method = NA_character_)

\S4method{smd}{numeric}(x, g, method = NA_character_)

\S4method{smd}{integer}(x, g, method = NA_character_)

\S4method{smd}{logical}(x, g, method = NA_character_)

\S4method{smd}{factor}(x, g, method = NA_character_)
}
\arguments{
\item{x}{a vector of values}

\item{g}{a vector}

\item{method}{not yet implemented}
}
\value{
the standardized mean differences between levels of \code{g}
for values of \code{x}
}
\description{
Compute Standardized Mean Difference
}
