% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd}
\alias{smd}
\alias{smd,character,ANY,missing-method}
\alias{smd,character,ANY,numeric-method}
\alias{smd,logical,ANY,missing-method}
\alias{smd,logical,ANY,numeric-method}
\alias{smd,matrix,ANY,missing-method}
\alias{smd,matrix,ANY,numeric-method}
\alias{smd,list,ANY,missing-method}
\alias{smd,list,ANY,numeric-method}
\alias{smd,data.frame,ANY,missing-method}
\alias{smd,data.frame,ANY,numeric-method}
\title{Compute Standardized Mean Difference}
\usage{
smd(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{character,ANY,missing}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{character,ANY,numeric}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{logical,ANY,missing}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{logical,ANY,numeric}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{matrix,ANY,missing}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{matrix,ANY,numeric}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{list,ANY,missing}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{list,ANY,numeric}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{data.frame,ANY,missing}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)

\S4method{smd}{data.frame,ANY,numeric}(x, g, w, std.error = FALSE, na.rm = FALSE, gref = 1L)
}
\arguments{
\item{x}{a vector of values}

\item{g}{a vector of groupings to compare}

\item{w}{a vector of weights (optional)}

\item{std.error}{Logical indicator for computing standard errors using
\code{\link{compute_smd_var}}.}

\item{na.rm}{remove \code{NA} values from \code{x}? Defaults to \code{FALSE}.}

\item{gref}{an integer indicating which level of \code{g} to use as the reference
group. Defaults to \code{1}.}
}
\value{
the standardized mean differences between levels of \code{g}
for values of \code{x}
}
\description{
Computes the standardized mean differnce (SMD) between two groups.
}
\examples{
x <- rnorm(100)
g <- rep(1:2, each = 50)
smd(x, g)
}
\seealso{
\code{\link{compute_smd}} for mathematical details
}
